hamro_pdf/hamro_django_backend/hamropdf_backend/requirements.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Database
mysqlclient==2.2.1
PyMySQL==1.1.0

# Authentication
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
PyJWT==2.8.0

# PDF Processing
PyPDF2==3.0.1
pypdf==3.17.4
pdf2image==1.16.3
Pillow==10.2.0
reportlab==4.0.8
pdfplumber==0.10.3
PyMuPDF==1.23.8
pikepdf==8.11.2
img2pdf==0.5.1
ocrmypdf==16.0.2

# Excel Processing  
openpyxl==3.1.2
xlsxwriter==3.1.9
pandas==2.1.4

# Document Conversion
python-docx==1.1.0
python-pptx==0.6.23

# File Handling
python-magic==0.4.27
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
uuid==1.30
gunicorn==21.2.0
whitenoise==6.6.0

# Image Processing
svglib==1.5.1
cairosvg==2.7.1
